requirements.txt
langchain==0.0.144
python-dotenv==1.0.0
termcolor==2.2.0
